Hacky fix for the SchematicSerializer
The SchematicSerializer is not adding the ConnectionTypes to the type
tree correctly which is causing NPEs when trying to serialize the
synthesized schematics used to test the backend. Temporarily change
the channel attributes in constraints to be "Real" instead of
"Connection" Type values.

This will break the following tests:
  TestChannelPlacementConstraintStrategy > testTwoNodes
  TestAnalyticalPressureFlowStrategy > testTwoNodes

When the serializer is fixed in mainline fix these tests and merge into
develop.